+++ src/authentication/rfl/authentication/ldap.py
@@ -26,6 +26,7 @@ class LDAPAuthentifier:
         user_class: str = "posixAccount",
         user_name_attribute: str = "uid",
         user_fullname_attribute: str = "cn",
+        user_primary_group_attribute: str = "gidNumber",
         group_name_attribute: str = "cn",
         cacert: Optional[Path] = None,
         starttls: bool = False,
@@ -40,6 +41,7 @@ class LDAPAuthentifier:
         self.group_base = group_base
         self.user_name_attribute = user_name_attribute
         self.user_fullname_attribute = user_fullname_attribute
+        self.user_primary_group_attribute = user_primary_group_attribute
         self.group_name_attribute = group_name_attribute
         self.starttls = starttls
         self.bind_dn = bind_dn
@@ -97,14 +99,14 @@ class LDAPAuthentifier:
     def _get_user_info(
         self, connection: ldap.ldapobject.LDAPObject, user_dn: str
     ) -> Tuple[str, int]:
-        """Return fullname and gidNumber from the provided user DN."""
+        """Return fullname and primary group number from the provided user DN."""
         search_filter = f"(objectClass={self.user_class})"
         try:
             results = connection.search_s(
                 user_dn,
                 ldap.SCOPE_BASE,
                 search_filter,
-                [self.user_fullname_attribute, "gidNumber"],
+                [self.user_fullname_attribute, self.user_primary_group_attribute],
             )
         except ldap.NO_SUCH_OBJECT as err:
             raise LDAPAuthenticationError(f"Unable to find user DN {user_dn}") from err
@@ -126,24 +128,24 @@ class LDAPAuthentifier:
                 "attribute from user entries"
             ) from err
         try:
-            gidNumber = int(results[0][1]["gidNumber"][0])
+            gid = int(results[0][1][self.user_primary_group_attribute][0])
         except KeyError as err:
             raise LDAPAuthenticationError(
-                "Unable to extract user primary group with gidNumber attribute from "
-                "user entries"
+                "Unable to extract user primary group with "
+                f"{self.user_primary_group_attribute} attribute from user entries"
             ) from err
-        return fullname, gidNumber
+        return fullname, gid
 
     def _get_groups(
         self,
         connection: ldap.ldapobject.LDAPObject,
         user_name: str,
         user_dn: str,
-        gidNumber: int,
+        gid: int,
     ) -> List[str]:
         """Return the list of groups whose provided user is member, including its
-        gidNumber. This function supports both RFC 2307 (aka. NIS schema) and RFC 2307
-        bis schema."""
+        primary group ID. This function supports both RFC 2307 (aka. NIS schema) and
+        RFC 2307bis schema."""
         # Standard RFC 2307 (aka. NIS) schema has group entries with posixGroup
         # structural class. Group members are declared with memberUid attributes (with
         # user cn as values).
@@ -152,12 +154,11 @@ class LDAPAuthentifier:
         # class and group members are declared with member attributes (with full user dn
         # as values).
         #
-        # In both cases, user primary group declared with gidNumber attribute in user
-        # entry must not be forgiven.
+        # In both cases, user primary group declared in user entry must not be forgiven.
         search_filter = (
             "(&"
             "(|(objectClass=posixGroup)(objectClass=groupOfNames))"
-            f"(|(memberUid={user_name})(member={user_dn})(gidNumber={gidNumber})))"
+            f"(|(memberUid={user_name})(member={user_dn})(gidNumber={gid})))"
         )
         try:
             results = connection.search_s(
@@ -180,7 +181,7 @@ class LDAPAuthentifier:
             logger.warning(
                 "Unable to find groups in LDAP for user %s or gidNumber %s",
                 user_name,
-                gidNumber,
+                gid,
             )
         try:
             return [
@@ -215,8 +216,8 @@ class LDAPAuthentifier:
             # Try simple authentication with user/password on LDAP directory
             user_dn = f"{self.user_name_attribute}={user},{self.user_base}"
             connection.simple_bind_s(user_dn, password)
-            fullname, gidNumber = self._get_user_info(connection, user_dn)
-            groups = self._get_groups(connection, user, user_dn, gidNumber)
+            fullname, gid = self._get_user_info(connection, user_dn)
+            groups = self._get_groups(connection, user, user_dn, gid)
         except ldap.SERVER_DOWN as err:
             raise LDAPAuthenticationError(
                 f"LDAP server {self.uri.geturl()} is unreachable"
@@ -294,10 +295,10 @@ class LDAPAuthentifier:
 
         try:
             for (user, user_dn) in self._list_user_dn(connection):
-                fullname, gidNumber = self._get_user_info(connection, user_dn)
+                fullname, gid = self._get_user_info(connection, user_dn)
                 groups = []
                 if with_groups:
-                    groups = self._get_groups(connection, user, user_dn, gidNumber)
+                    groups = self._get_groups(connection, user, user_dn, gid)
                     # Skip the user if not member of any of the restricted groups
                     if not self._in_restricted_groups(groups):
                         logger.debug(